Move http method wrappers to seperate file
